+++ CHANGELOG.md
@@ -19,3 +19,12 @@
 - Moved library JavaScript files to assets/lib
 
 ---
+
+### 0.2.1
+
+- Removed DiscordLogin function and replaced it with GetTokenData function  
+  *So the socket-io library is nor pulled out of the Discord functions which makes it a bit more abstract for other usages which could need this data maybe laster.*
+- Added GetTokenValidity function to simply check if a token is valid or not
+- Added a lot of comments for better understanding and easier contribution
+
+---